style: :art: folder restructuring for optimization
3
